requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
certifi==2025.11.12
cloudevents==1.12.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
openai==2.12.0
packaging==25.0
pika==1.3.2
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
# Ready-Made First: Agent frameworks (Этап 4)
litellm==1.80.10
langgraph==1.0.5
